Restructure the test. Make test run only when profile test is active to ease dev
