+++ src/test/java/org/craftercms/studio/testing/it/ui/SiteAvailabilityTest.java
@@ -1,7 +1,7 @@
-package org.craftercms.studio.testing.ui;
+package org.craftercms.studio.testing.it.ui;
 
 import org.testng.annotations.Test;
-import org.craftercms.studio.testing.base.ui.BaseSeleniumTest;
+import org.craftercms.studio.testing.it.base.ui.BaseSeleniumTest;
 
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNotSame;
