+++ src/test/resources/ui.xml
@@ -3,13 +3,13 @@
 
     <test name="Selenium Integrity">
         <packages>
-            <package name="org.sample.ui"/>
+            <package name="org.craftercms.studio.testing.it.ui"/>
         </packages>
     </test>
 
     <test name="Selenium Create site">
         <packages>
-            <package name="org.sample.ui.site"/>
+            <package name="org.craftercms.studio.testing.it.ui.site"/>
         </packages>
     </test>
